+++ src/services/BlockRepository.ts
@@ -3,7 +3,7 @@ import {injectable} from 'inversify';
 import {HexStringWith0x} from '../types/custom';
 import {v4 as uuid} from 'uuid';
 import Block from '../models/Block';
-import Leaf from '../models/Leaf';
+import Leaf from '../types/Leaf';
 import {BigNumber} from 'ethers';
 import {SignedBlockConsensus} from '../types/Consensus';
 
@@ -43,29 +43,12 @@ class BlockRepository {
     block.root = params.root;
     block.data = this.treeDataFor(params.leaves);
     block.fcdKeys = params.fcdKeys;
-    await this.attachLeavesToBlock(params.leaves, params.blockId);
+
     return getModelForClass(Block).findOneAndUpdate({blockId: block.blockId}, block, {
       upsert: true,
       setDefaultsOnInsert: true,
       new: true,
     });
-  }
-
-  private async attachLeavesToBlock(leaves: Leaf[], blockId: number): Promise<void> {
-    return getModelForClass(Leaf)
-      .updateMany(
-        {
-          _id: {
-            $in: leaves.map((leaf) => leaf._id),
-          },
-        },
-        {
-          $set: {
-            blockId: blockId,
-          },
-        },
-      )
-      .exec();
   }
 
   private treeDataFor(leaves: Leaf[]): Record<string, HexStringWith0x> {